+++ packages/ui-vue/src/index.ts
@@ -8,3 +8,5 @@ export { default as PVList } from "./PVList.vue";
 export { default as SearchLog } from "./SearchLog.vue";
 export { default as ScoreBadge } from "./ScoreBadge.vue";
 export { default as BoardComponents } from "./BoardComponents.vue";
+export { default as ChessBoard } from "./ChessBoard.vue";
+export { default as ShogiBoard } from "./ShogiBoard.vue";